Move process runner to the right runner package
