+++ core/src/main/scala/stryker4s/config/Config.scala
@@ -1,13 +1,11 @@
 package stryker4s.config
 
 import better.files._
-import org.apache.logging.log4j.Level
 import pureconfig.ConfigWriter
 import stryker4s.run.report.{ConsoleReporter, MutantRunReporter}
 
 case class Config(mutate: Seq[String] = Seq("**/main/scala/**/*.scala"),
                   baseDir: File = File.currentWorkingDirectory,
-                  testRunner: TestRunner = CommandRunner("sbt", "test"),
                   reporters: List[MutantRunReporter] = List(new ConsoleReporter),
                   files: Option[Seq[String]] = None,
                   excludedMutations: ExcludedMutations = ExcludedMutations(),
