+++ core/src/test/resources/stryker4sconfs/filled.conf
@@ -7,12 +7,6 @@ stryker4s {
 
   base-dir: "/tmp/project"
 
-  test-runner: {
-    type=commandrunner
-    command: "mvn"
-    args: "clean test"
-  }
-
   reporters: ["console"]
 
   excluded-mutations: ["BooleanLiteral"]
