+++ core/src/test/resources/stryker4sconfs/wrongReporter.conf
@@ -7,11 +7,5 @@ stryker4s {
 
   base-dir: "/tmp/project"
 
-  test-runner: {
-    type=commandrunner
-    command: "mvn"
-    args: "clean test"
-  }
-
   reporters: ["dsadsa"]
 }
