--- core/src/test/resources/stryker4sconfs/wrongTestRunner.conf
@@ -1,7 +1,0 @@
-stryker4s {
-  test-runner: {
-    type=someOtherTestRunner
-    command=bar
-    args=foo
-  }
-}
