+++ core/src/test/scala/stryker4s/config/ConfigReaderTest.scala
@@ -17,7 +17,6 @@ class ConfigReaderTest extends Stryker4sSuite with LogMatchers {
 
       result.baseDir shouldBe File.currentWorkingDirectory
       result.mutate shouldBe Seq("**/main/scala/**/*.scala")
-      result.testRunner shouldBe an[CommandRunner]
       result.reporters.head shouldBe an[ConsoleReporter]
     }
 
@@ -46,17 +45,16 @@ class ConfigReaderTest extends Stryker4sSuite with LogMatchers {
 
       result.baseDir shouldBe File("/tmp/project")
       result.mutate shouldBe Seq("bar/src/main/**/*.scala", "foo/src/main/**/*.scala", "!excluded/file.scala")
-      result.testRunner shouldBe an[CommandRunner]
       result.reporters.head shouldBe an[ConsoleReporter]
       result.excludedMutations shouldBe ExcludedMutations(Set("BooleanLiteral"))
     }
 
-    it("should return a failure on an invalid exclusion mutator") {
-      val confPath = FileUtil.getResource("stryker4sconfs/wrongTestRunner.conf")
-
-      lazy val result = ConfigReader.readConfig(confPath)
-      val exc = the[ConfigReaderException[_]] thrownBy result
-    }
+//    it("should return a failure on an invalid exclusion mutator") {
+//      val confPath = FileUtil.getResource("stryker4sconfs/wrongTestRunner.conf")
+//
+//      lazy val result = ConfigReader.readConfig(confPath)
+//      val exc = the[ConfigReaderException[_]] thrownBy result
+//    }
 
     it("should return a failure on a misshapen test runner") {
       val confPath = FileUtil.getResource("stryker4sconfs/invalidExcludedMutation.conf")
