+++ runners/command-runner/src/main/scala/stryker4s/config/ProcessRunnerConfig.scala
@@ -1,0 +1,5 @@
+package stryker4s.config
+
+import stryker4s.run.process.Command
+
+case class ProcessRunnerConfig(testRunnerCommand: Command)
