+++ runners/command-runner/src/main/scala/stryker4s/run/ProcessMutantRunner.scala
@@ -1,4 +1,4 @@
-package stryker4s.run.process
+package stryker4s.run
 
 import java.nio.file.Path
 
@@ -6,7 +6,7 @@ import better.files.File
 import stryker4s.config.Config
 import stryker4s.model._
 import stryker4s.mutants.findmutants.SourceCollector
-import stryker4s.run.MutantRunner
+import stryker4s.run.process.{Command, ProcessRunner}
 
 import scala.concurrent.TimeoutException
 import scala.util.{Failure, Success}
